+++ synapse-edison/src/test/java/de/otto/synapse/edison/health/StartupHealthIndicatorTest.java
@@ -1,11 +1,14 @@
 package de.otto.synapse.edison.health;
 
+import com.google.common.collect.ImmutableMap;
 import de.otto.synapse.eventsource.EventSource;
 import de.otto.synapse.eventsource.EventSourceNotification;
 import org.junit.Test;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.Status;
 
+import static de.otto.synapse.eventsource.EventSourceNotification.builder;
+import static java.util.Arrays.asList;
 import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -17,7 +20,10 @@ public class StartupHealthIndicatorTest {
     @Test
     public void shouldInitiallyIndicateDown() {
         // given
-        StartupHealthIndicator healthCheck = new StartupHealthIndicator();
+        HealthProperties properties = new HealthProperties();
+        properties.setChannels(asList("some-stream", "other-stream"));
+
+        StartupHealthIndicator healthCheck = new StartupHealthIndicator(properties);
 
         // when
         Health health = healthCheck.health();
@@ -27,26 +33,57 @@ public class StartupHealthIndicatorTest {
     }
 
     @Test
-    public void shouldIndicateUpWhenFirstEventSourceIsFinished() {
+    public void shouldIndicateDownBeforeLastChannelHasFinished() {
         // given
-        StartupHealthIndicator healthCheck = new StartupHealthIndicator();
+        HealthProperties properties = new HealthProperties();
+        properties.setChannels(asList("some-stream", "other-stream"));
+        StartupHealthIndicator healthCheck = new StartupHealthIndicator(properties);
 
         EventSource mockEventSource = mock(EventSource.class);
         when(mockEventSource.getChannelName()).thenReturn("some-stream");
 
-        EventSourceNotification eventSourceNotification = EventSourceNotification.builder()
-                .withStatus(EventSourceNotification.Status.FINISHED)
-                .withMessage("some message")
-                .withChannelName("some-stream")
-                .build();
-        healthCheck.onEventSourceNotification(eventSourceNotification);
-
         // when
-        Health health = healthCheck.health();
+        healthCheck.onEventSourceNotification(builder()
+                .withStatus(EventSourceNotification.Status.FINISHED)
+                .withMessage("Finished some-stream")
+                .withChannelName("some-stream")
+                .build());
 
         // then
+        Health health = healthCheck.health();
+        assertThat(health.getStatus(), is(Status.DOWN));
+        assertThat(health.getDetails(), hasEntry("some-stream", ImmutableMap.of("message", "Finished some-stream", "status", "FINISHED")));
+        assertThat(health.getDetails(), hasEntry("other-stream", ImmutableMap.of("message", "Channel not yet finished", "status", "NOT FINISHED")));
+
+    }
+
+    @Test
+    public void shouldIndicateUpAfterLastChannelHasFinished() {
+        // given
+        HealthProperties properties = new HealthProperties();
+        properties.setChannels(asList("some-stream", "other-stream"));
+        StartupHealthIndicator healthCheck = new StartupHealthIndicator(properties);
+
+        EventSource mockEventSource = mock(EventSource.class);
+        when(mockEventSource.getChannelName()).thenReturn("some-stream");
+
+        // when
+        healthCheck.onEventSourceNotification(builder()
+                .withStatus(EventSourceNotification.Status.FINISHED)
+                .withMessage("Finished some-stream")
+                .withChannelName("some-stream")
+                .build());
+
+        healthCheck.onEventSourceNotification(builder()
+                .withStatus(EventSourceNotification.Status.FINISHED)
+                .withMessage("Finished other-stream")
+                .withChannelName("other-stream")
+                .build());
+
+        // then
+        Health health = healthCheck.health();
         assertThat(health.getStatus(), is(Status.UP));
-        assertThat(health.getDetails(), hasEntry("message", "some message"));
-        assertThat(health.getDetails(), hasEntry("stream", "some-stream"));
+        assertThat(health.getDetails(), hasEntry("some-stream", ImmutableMap.of("message", "Finished some-stream", "status", "FINISHED")));
+        assertThat(health.getDetails(), hasEntry("other-stream", ImmutableMap.of("message", "Finished other-stream", "status", "FINISHED")));
     }
 }